feat: reuse `User` type across all requests
